55: extract `<PrivateRoute>`, `<Register>` into the newly-created `frontend/src/features/auth/` directory + implement a `requestHandlers` object within the newly-created `frontend/src/testHelpers.ts`
